+++ app/models/__init__.py
@@ -4,4 +4,5 @@ from .exercise import Exercise
 from .workout import Workout 
 from .event import Event
 from .comment import Comment
+from .workout_exercise import WorkoutExercise
 from .db import environment, SCHEMA
